+++ lib/kill.go
@@ -1,24 +1,10 @@
 package lib
 
 import (
-	"github.com/docker/docker/pkg/signal"
 	"github.com/kubernetes-sigs/cri-o/oci"
-	"github.com/kubernetes-sigs/cri-o/utils"
 	"github.com/pkg/errors"
-	"os"
 	"syscall"
 )
-
-// Reverse lookup signal string from its map
-func findStringInSignalMap(killSignal syscall.Signal) (string, error) {
-	for k, v := range signal.SignalMap {
-		if v == killSignal {
-			return k, nil
-		}
-	}
-	return "", errors.Errorf("unable to convert signal to string")
-
-}
 
 // ContainerKill sends the user provided signal to the containers primary process.
 func (c *ContainerServer) ContainerKill(container string, killSignal syscall.Signal) (string, error) { // nolint
@@ -33,17 +19,11 @@ func (c *ContainerServer) ContainerKill(container string, killSignal syscall.Sig
 	if cStatus.Status != oci.ContainerStateRunning {
 		return "", errors.Errorf("cannot kill container %s: it is not running", container)
 	}
-	signalString, err := findStringInSignalMap(killSignal)
-	if err != nil {
+
+	if err = c.runtime.SignalContainer(ctr, killSignal); err != nil {
 		return "", err
 	}
-	rPath, err := c.runtime.Path(ctr)
-	if err != nil {
-		return "", err
-	}
-	if err := utils.ExecCmdWithStdStreams(os.Stdin, os.Stdout, os.Stderr, rPath, "kill", ctr.ID(), signalString); err != nil {
-		return "", err
-	}
+
 	c.ContainerStateToDisk(ctr)
 	return ctr.ID(), nil
 }
